# doc/openAPI/eem_swagger.yaml
swagger: '2.0'
info:
  description: 5G EVE Experiment Execution Management ReST API
  version: 1.0.1
  title: Experiment Execution Management API
  termsOfService: 'https://www.5g-eve.eu/'
  contact: {}
  license:
    name: Apache2
    url: 'http://unlicense.org'
host: 'localhost:8080'
basePath: /
...
paths:
  /eem/:
    get:
      tags:
        - EEM Application
      summary: List API versions
      operationId: listVersionsv1
      produces:
        - application/json
      responses:
        '200':
          description: 200 response
          schema:
            type: object
        '300':
          description: 300 response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    options:
      tags:
        - EEM Application
      summary: eemOptions
      operationId: eemOptions
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: The Option operation is needed for compatibility reasons with JS applications
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /eem/experiment_executions:
    get:
      tags:
        - EEM Operations
      summary: List all experiments available
      operationId: eemExperimentExecutionsGet
      produces:
        - application/json
      parameters:
        - name: state
          in: query
          description: Execution state of the experiment
          required: false
          type: string
          allowEmptyValue: false
          enum:
            - INIT
            - CONFIGURING
            - RUNNING
            - RUNNING_STEP
            - PAUSED
            - VALIDATING
            - COMPLETED
            - ABORTING
            - ABORTED
            - FAILED
      responses:
        '200':
          description: List of all experiments
          schema:
            type: array
            items:
              $ref: '#/definitions/ExperimentExecution'
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Not Found
        '412':
          description: Precondition Failed
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
    post:
      tags:
        - EEM Operations
      summary: Requests a new experiment execution
      operationId: eemExperimentExecutionsPost
      consumes:
        - application/json
      produces:
        - application/json
        - text/plain
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Experiment execution identifier
          schema:
            type: string
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Not Found
        '412':
          description: Precondition Failed
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
    options:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsOptions
      operationId: eemExperimentExecutionsOptions
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: The Option operation is needed for compatibility reasons with JS     applications
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  '/eem/experiment_executions/{id}':
    get:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdGet
      description: API to get experiment execution details
      operationId: eemExperimentExecutionsIdGet
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Return experiment execution details
          schema:
            $ref: '#/definitions/ExperimentExecution'
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Identifier of the experiment execution not found
          schema:
            $ref: '#/definitions/ErrorInfo'
        '412':
          description: Precondition Failed
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
    delete:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdDelete
      operationId: eemExperimentExecutionsIdDelete
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: Return 204 No content upon action completion
          schema:
            type: object
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Identifier of the Experiment execution not found
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
    options:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdOptions
      operationId: eemExperimentExecutionsIdOptions
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Return 200 OK upon action completion
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  '/eem/experiment_executions/{id}/abort':
    post:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdAbortPost
      operationId: eemExperimentExecutionsIdAbortPost
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Return 200 OK upon action completion
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Identifier of the experiment execution not found
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
  '/eem/experiment_executions/{id}/pause':
    post:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdPausePost
      operationId: eemExperimentExecutionsIdPausePost
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Return 200 OK upon action completion
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Identifier of the experiment execution not found
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
  '/eem/experiment_executions/{id}/resume':
    post:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdResumePost
      operationId: eemExperimentExecutionsIdResumePost
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Return 200 OK upon action completion
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Identifier of the experiment execution not found
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
  '/eem/experiment_executions/{id}/run':
    post:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdRunPost
      operationId: eemExperimentExecutionsIdRunPost
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: body
          required: true
          schema:
            $ref: '#/definitions/ExperimentExecutionRequest'
        - name: id
          in: path
          description: id
          required: true
          type: string
        - name: runType
          in: query
          description: 'Determine the type of run. If not present, the default value is RUN_ALL'
          required: false
          type: string
          allowEmptyValue: false
          enum:
            - RUN_IN_STEPS
            - RUN_ALL
      responses:
        '200':
          description: Return 200 OK upon action completion
          schema:
            type: object
        '201':
          description: Created
        '400':
          description: Malformatted request
          schema:
            $ref: '#/definitions/ErrorInfo'
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Identifier of the experiment execution not found
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
  '/eem/experiment_executions/{id}/step':
    post:
      tags:
        - EEM Operations
      summary: eemExperimentExecutionsIdStepPost
      operationId: eemExperimentExecutionsIdStepPost
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Return 200 OK upon action completion
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Operation not authorised
          schema:
            $ref: '#/definitions/ErrorInfo'
        '403':
          description: Operation forbidden
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: Identifier of the experiment execution not found
          schema:
            $ref: '#/definitions/ErrorInfo'
        '500':
          description: Operation failed
          schema:
            $ref: '#/definitions/ErrorInfo'
      deprecated: false
...
definitions:
...
  ErrorInfo:
    type: object
    required:
      - detail
      - status
    properties:
      detail:
        type: string
        description: 'A human-readable explanation specific to this occurrence of the problem. '
      status:
        type: integer
        format: int32
        description: 'The HTTP status code for this occurrence of the problem. The HTTP status code ([RFC7231], Section 6) generated by the origin server for this occurrence of the problem. '
    title: ErrorInfo
    description: 'The definition of the general "ProblemDetails" data structure from IETF RFC 7807 [19] is reproduced in this structure.  '
  ExecutionResult:
    type: object
    required:
      - result
    properties:
      result:
        type: string
    title: ExecutionResult
  ExperimentExecution:
    type: object
    required:
      - executionId
      - state
    properties:
      errorMessage:
        type: string
        description: Error Message
      executionId:
        type: string
        description: Identifier of the executed experiment
      experimentDescriptorId:
        type: string
      nsInstanceId:
        type: string
      reportUrl:
        type: string
        description: URL containing the report of the execution
      state:
        type: string
        enum:
          - INIT
          - CONFIGURING
          - RUNNING
          - RUNNING_STEP
          - PAUSED
          - VALIDATING
          - COMPLETED
          - ABORTING
          - ABORTED
          - FAILED
      testCaseDescriptorConfiguration:
        type: array
        description: Test cases for the requested execution
        items:
          $ref: '#/definitions/TestCaseExecutionConfiguration'
      testCaseResult:
        type: object
        additionalProperties:
          $ref: '#/definitions/ExecutionResult'
    title: ExperimentExecution
  ExperimentExecutionRequest:
    type: object
    required:
      - executionId
      - experimentDescriptorId
      - nsInstanceId
    properties:
      executionId:
        type: string
      experimentDescriptorId:
        type: string
      nsInstanceId:
        type: string
      testCaseDescriptorConfiguration:
        type: object
        additionalProperties:
          type: object
          additionalProperties:
            type: string
    title: ExperimentExecutionRequest
...
  TestCaseExecutionConfiguration:
    type: object
    properties:
      execConfiguration:
        type: object
        additionalProperties:
          type: string
      tcDescriptorId:
        type: string
    title: TestCaseExecutionConfiguration